# DenseSet::Grow optimization