fixes: #3870

Before BM_Grow        721975 ns       721914 ns         1020
Now    BM_Grow        598066 ns       597285 ns         1237