@romange 2024-10-25
And I still think that cached links is unnecessary over complication.