@BorysTheDev 2024-10-25
> And I still think that cached links is unnecessary over complication.

We already use FreeLink and NewLink, we can create a small class that incapsulate this logic and cached links. it's very smooth and really good for performance even for ADD/REM operations